+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 
 ## 2026-04-24
+- `14:45` -- **Added** `apps/desktopProbe/tests/quietHours/schedule.test.ts`
 - `14:45` -- **Updated** `apps/desktopProbe/src/lib/types.ts`
 - `14:41` -- **Updated** `apps/desktopProbe/package.json`
 - `14:40` -- **Added** `apps/desktopProbe/jest.config.ts`